// backend/src/modules/production-orders/dto/update-production-order-disinfectant-preparation.dto.ts
export class UpdateProductionOrderDisinfectantPreparationDto {
  workshop_id?: number | string | null;
  disinfectant_name?: string | null;
  purpose?: string | null;
  base_material_name?: string | null;
  base_material_content?: number | string | null;
  base_material_amount_l?: number | string | null;
  prepared_volume_l?: number | string | null;
  actual_concentration?: number | string | null;
}